Refactor: moved inner classes from AnnotationEmitter to their related files
